functionality graph.describe
